create IAsyncRepository and EfRepository

+++ ApplicationCore/Models/MovieCardResponseModel.cs
@@ -11,6 +11,6 @@ namespace ApplicationCore.Models
         public int Id { get; set; }
         public string Title { get; set; }
         public string PostedUrl { get; set; }
-        public int Budget { get; set; }
+        public decimal Budget { get; set; }
     }
 }
